requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
ansible==8.7.0
ansible-core==2.15.12
anyio==4.0.0
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto==2.49.0
botocore==1.29.43
bs4==0.0.2
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.0
cryptography==43.0.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.3
django-storages-redux==1.3.3
dnspython==2.4.2
edge-tts==6.1.11
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.0
filelock==3.12.3
Flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gensim==3.8.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.0.7
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-text-splitters==0.2.2
langsmith==0.1.83
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mpv==1.0.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
orjson==3.10.6
outcome==1.2.0
packaging==24.1
pandas==2.1.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pgvector==0.2.1
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.16.1
pymongo==4.4.1
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resolvelib==1.0.1
s3transfer==0.6.0
safetensors==0.3.3
scikit-learn==1.3.0
scikit-posthocs==0.7.0
scipy==1.11.2
seaborn==0.12.2
selenium==4.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.13.5
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.1
trio==0.22.0
trio-websocket==0.9.2
trueskill==0.4.5
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.20
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
webdriver-manager==4.0.2
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
yellowbrick==1.5
zipp==3.16.2
